refactor(signup): extract error in own file
