+++ src/app/visitor/signup/domain/email-already-taken.error.ts
@@ -1,0 +1,6 @@
+export class EmailAlreadyTakenError extends Error {
+  constructor(readonly email: string) {
+    super(`Email ${email} is already taken. Please try another email.`);
+    this.name = 'EmailAlreadyTakenError';
+  }
+}
