+++ src/app/visitor/signup/signup.page.component.ts
@@ -4,7 +4,7 @@ import { UserStore } from '../../core/store/user.store';
 import { Visitor } from '../../core/entity/user.interface';
 import { RegisterUserUseCase } from './domain/register-user.use-case';
 import { Router } from '@angular/router';
-import { EmailAlreadyTakenError } from 'src/app/core/port/authentication.service';
+import { EmailAlreadyTakenError } from './domain/email-already-taken.error';
 
 @Component({
   standalone: true,
